Got most of the addnewsongs page working, and moved it to the pages folder
