+++ frontend/src/pages/AddNewSongs/AddNewSongsStyles.jsx
@@ -2,7 +2,13 @@ import styled from "styled-components";
 
 const AddNewSongsStyles = styled.div`
   display: flex;
-  color: white;
+  flex-direction: column;
+  ${'' /* color: white; */}
+
+  .song-list {
+    margin: 0 auto;
+    width: 80%;
+  }
 `
 
 export default AddNewSongsStyles;
